+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Change Log
 
 - added dill serialization support (https://pypi.python.org/pypi/dill)
 - fixed dotted attribute client code in the ``attributes`` example
+- handles EINTR signal and will continue the server loop now in this case, on Python 3.4 and newer.
 
 
 **Pyro 4.41**